backend/requirements.txt
blinker==1.6.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
Flask==2.2.5
Flask-Cors==4.0.0
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
greenlet==3.0.3
idna==3.6
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.19.0
midtransclient==1.4.2
packaging==23.2
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==0.21.1
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.7.1
urllib3==1.26.6
Werkzeug==2.2.3
zipp==3.15.0